refactor: move client tests to source file
